src/test/java/guru/springframework/services/ingredient/IngredientServiceImpTest.java
package guru.springframework.services.ingredient;

import guru.springframework.converters.ingredients.IngredientCommandToIngredient;
import guru.springframework.converters.ingredients.IngredientToIngredientCommand;
import guru.springframework.domian.Ingredient;
import guru.springframework.repositories.IngredientRepository;
import guru.springframework.repositories.RecipeRepository;
import guru.springframework.repositories.UnitOfMeasureRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.Optional;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.Mockito.when;

class IngredientServiceImpTest {
    @Mock
    IngredientRepository ingredientRepository;
    @Mock
    private RecipeRepository recipeRepository;
    @Mock
    private UnitOfMeasureRepository unitOfMeasureRepository;
    IngredientToIngredientCommand ingredientToIngredientCommand;

    IngredientCommandToIngredient ingredientCommandToIngredient;

    IngredientService ingredientService;
    String ID_VALUE = "1";

    @BeforeEach
    void setUp() {
        MockitoAnnotations.initMocks(this);
        ingredientService = new IngredientServiceImp(ingredientRepository, ingredientToIngredientCommand, ingredientCommandToIngredient, recipeRepository, unitOfMeasureRepository);
    }

    @Test
    void findCommandById() {
        Ingredient ingredient = new Ingredient();
        ingredient.setId(ID_VALUE);
        Optional<Ingredient> ingredientData = Optional.of(ingredient);
        when(ingredientRepository.findById(anyLong())).thenReturn(ingredientData);
        Ingredient ingredientReturn = ingredientService.findById(new Long(ID_VALUE));
        assertNotNull(ingredientReturn);
    }
}